inference 수정
