+++ dataset/XRayDataset.py
@@ -10,23 +10,7 @@ from sklearn.model_selection import GroupKFold
 import torch
 from torch.utils.data import Dataset
 
-
-IMAGE_ROOT = "./data/train/DCM"
-LABEL_ROOT = "./data/train/outputs_json"
-
-
-CLASSES = [
-    'finger-1', 'finger-2', 'finger-3', 'finger-4', 'finger-5',
-    'finger-6', 'finger-7', 'finger-8', 'finger-9', 'finger-10',
-    'finger-11', 'finger-12', 'finger-13', 'finger-14', 'finger-15',
-    'finger-16', 'finger-17', 'finger-18', 'finger-19', 'Trapezium',
-    'Trapezoid', 'Capitate', 'Hamate', 'Scaphoid', 'Lunate',
-    'Triquetrum', 'Pisiform', 'Radius', 'Ulna',
-]
-
-CLASS2IND = {v: i for i, v in enumerate(CLASSES)}
-
-IND2CLASS = {v: k for k, v in CLASS2IND.items()}
+from configs.config import IMAGE_ROOT, LABEL_ROOT, CLASSES, CLASS2IND
 
 pngs = {
     os.path.relpath(os.path.join(root, fname), start=IMAGE_ROOT)
@@ -213,3 +197,31 @@ class XRayDataset(Dataset):
             cv2.fillPoly(class_label, [points], 1)
             label[..., class_ind] = class_label
         return image, label
+
+class XRayInferenceDataset(Dataset):
+    def __init__(self, transforms=None):
+        _filenames = pngs
+        _filenames = np.array(sorted(_filenames))
+
+        self.filenames = _filenames
+        self.transforms = transforms
+
+    def __len__(self):
+        return len(self.filenames)
+
+    def __getitem__(self, item):
+        image_name = self.filenames[item]
+        image_path = os.path.join(IMAGE_ROOT, image_name)
+
+        image = cv2.imread(image_path)
+        image = image / 255.
+
+        if self.transforms is not None:
+            inputs = {"image": image}
+            result = self.transforms(**inputs)
+            image = result["image"]
+
+        image = image.transpose(2, 0, 1)
+        image = torch.from_numpy(image).float()
+
+        return image, image_name
