+++ dataset/tta.py
@@ -1,0 +1,60 @@
+import numpy as np
+import ttach as tta
+from tqdm.auto import tqdm
+import torch
+import torch.nn.functional as F
+from configs.config import CLASSES, IND2CLASS
+
+def encode_mask_to_rle(mask):
+    pixels = mask.flatten()
+    pixels = np.concatenate([[0], pixels, [0]])
+    runs = np.where(pixels[1:] != pixels[:-1])[0] + 1
+    runs[1::2] -= runs[::2]
+    return ' '.join(str(x) for x in runs)
+
+
+def decode_rle_to_mask(rle, height, width):
+    s = rle.split()
+    starts, lengths = [np.asarray(x, dtype=int) for x in (s[0:][::2], s[1:][::2])]
+    starts -= 1
+    ends = starts + lengths
+    img = np.zeros(height * width, dtype=np.uint8)
+
+    for lo, hi in zip(starts, ends):
+        img[lo:hi] = 1
+
+    return img.reshape(height, width)
+
+
+def test_with_tta(model, data_loader, thr=0.6):
+    model = model.cuda()
+    model.eval()
+
+    # TTA 정의
+    tta_transforms = tta.Compose(
+        [
+            tta.HorizontalFlip(),
+        ]
+    )
+    tta_model = tta.SegmentationTTAWrapper(model, tta_transforms)
+
+    rles = []
+    filename_and_class = []
+    with torch.no_grad():
+        n_class = len(CLASSES)
+
+        for step, (images, image_names) in tqdm(enumerate(data_loader), total=len(data_loader)):
+            images = images.cuda()
+            outputs = tta_model(images)
+
+            outputs = F.interpolate(outputs, size=(2048, 2048), mode="bilinear")
+            outputs = torch.sigmoid(outputs)
+            outputs = (outputs > thr).detach().cpu().numpy()
+
+            for output, image_name in zip(outputs, image_names):
+                for c, segm in enumerate(output):
+                    rle = encode_mask_to_rle(segm)
+                    rles.append(rle)
+                    filename_and_class.append(f"{IND2CLASS[c]}_{image_name}")
+
+    return rles, filename_and_class
