+++ inference/inference.py
@@ -15,154 +15,72 @@ import torch.nn.functional as F
 from torch.utils.data import Dataset, DataLoader
 import segmentation_models_pytorch as smp
 
-
-CLASSES = [
-    'finger-1', 'finger-2', 'finger-3', 'finger-4', 'finger-5',
-    'finger-6', 'finger-7', 'finger-8', 'finger-9', 'finger-10',
-    'finger-11', 'finger-12', 'finger-13', 'finger-14', 'finger-15',
-    'finger-16', 'finger-17', 'finger-18', 'finger-19', 'Trapezium',
-    'Trapezoid', 'Capitate', 'Hamate', 'Scaphoid', 'Lunate',
-    'Triquetrum', 'Pisiform', 'Radius', 'Ulna',
-]
-CLASS2IND = {v: i for i, v in enumerate(CLASSES)}
-IND2CLASS = {v: k for k, v in CLASS2IND.items()}
-SAVED_DIR = "checkpoints/result_unet/"
-IMAGE_ROOT = "./data/test/"
-
-model = smp.UnetPlusPlus(
-    encoder_name="efficientnet-b7",
-    encoder_weights="imagenet",
-    in_channels=3,
-    classes=29
-)
-
-model = torch.load(os.path.join(SAVED_DIR, "unetPP_base.pt"))
+from dataset.XRayDataset import *
+from configs import config as cfg
+from models import *
+from dataset.tta import *
 
 
-pngs = {
-    os.path.relpath(os.path.join(root, fname), start=IMAGE_ROOT)
-    for root, _dirs, files in os.walk(IMAGE_ROOT)
-    for fname in files
-    if os.path.splitext(fname)[1].lower() == ".png"
-}
+def main():
+    # model 불러오기
+    if cfg.MODEL == 'unetpp':
+        model = smp.UnetPlusPlus(
+            encoder_name="efficientnet-b7",
+            encoder_weights="imagenet",
+            in_channels=3,
+            classes=len(CLASSES),
+        ).cuda()
+    elif cfg.MODEL == 'conv':
+        model = UperNet_ConvNext_xlarge(num_classes=len(CLASSES))
+    elif cfg.MODEL == 'segformer':
+        model = SegFormer_B0(num_classes=len(CLASSES))
+    elif cfg.MODEL == 'deeplap':
+        model = DeepLabV3p(in_channels=3, num_classes=len(CLASSES))
+    else:
+        print("model not found.")
+        exit(0)
+
+    model = torch.load(os.path.join(cfg.SAVED_DIR, f"{cfg.PT_NAME}"))
 
 
-def encode_mask_to_rle(mask):
-    pixels = mask.flatten()
-    pixels = np.concatenate([[0], pixels, [0]])
-    runs = np.where(pixels[1:] != pixels[:-1])[0] + 1
-    runs[1::2] -= runs[::2]
-    return ' '.join(str(x) for x in runs)
+    # 데이터 불러오기
+    tf = A.Resize(1024, 1024)
 
+    test_dataset = XRayInferenceDataset(transforms=tf)
 
-def decode_rle_to_mask(rle, height, width):
-    s = rle.split()
-    starts, lengths = [np.asarray(x, dtype=int) for x in (s[0:][::2], s[1:][::2])]
-    starts -= 1
-    ends = starts + lengths
-    img = np.zeros(height * width, dtype=np.uint8)
-
-    for lo, hi in zip(starts, ends):
-        img[lo:hi] = 1
-
-    return img.reshape(height, width)
-
-
-class XRayInferenceDataset(Dataset):
-    def __init__(self, transforms=None):
-        _filenames = pngs
-        _filenames = np.array(sorted(_filenames))
-
-        self.filenames = _filenames
-        self.transforms = transforms
-
-    def __len__(self):
-        return len(self.filenames)
-
-    def __getitem__(self, item):
-        image_name = self.filenames[item]
-        image_path = os.path.join(IMAGE_ROOT, image_name)
-
-        image = cv2.imread(image_path)
-        image = image / 255.
-
-        if self.transforms is not None:
-            inputs = {"image": image}
-            result = self.transforms(**inputs)
-            image = result["image"]
-
-        image = image.transpose(2, 0, 1)
-        image = torch.from_numpy(image).float()
-
-        return image, image_name
-
-
-def test_with_tta(model, data_loader, thr=0.6):
-    model = model.cuda()
-    model.eval()
-
-    # TTA 정의
-    tta_transforms = tta.Compose(
-        [
-            tta.HorizontalFlip(),
-        ]
+    test_loader = DataLoader(
+        dataset=test_dataset,
+        batch_size=2,
+        shuffle=False,
+        num_workers=2,
+        drop_last=False
     )
-    tta_model = tta.SegmentationTTAWrapper(model, tta_transforms)
 
-    rles = []
-    filename_and_class = []
-    with torch.no_grad():
-        n_class = len(CLASSES)
-
-        for step, (images, image_names) in tqdm(enumerate(data_loader), total=len(data_loader)):
-            images = images.cuda()
-            outputs = tta_model(images)
-
-            outputs = F.interpolate(outputs, size=(2048, 2048), mode="bilinear")
-            outputs = torch.sigmoid(outputs)
-            outputs = (outputs > thr).detach().cpu().numpy()
-
-            for output, image_name in zip(outputs, image_names):
-                for c, segm in enumerate(output):
-                    rle = encode_mask_to_rle(segm)
-                    rles.append(rle)
-                    filename_and_class.append(f"{IND2CLASS[c]}_{image_name}")
-
-    return rles, filename_and_class
+    rles, filename_and_class = test_with_tta(model, test_loader)
 
 
-tf = A.Resize(1024, 1024)
+    preds = []
+    for rle in rles[:len(CLASSES)]:
+        pred = decode_rle_to_mask(rle, height=2048, width=2048)
+        preds.append(pred)
 
-test_dataset = XRayInferenceDataset(transforms=tf)
-
-test_loader = DataLoader(
-    dataset=test_dataset,
-    batch_size=2,
-    shuffle=False,
-    num_workers=2,
-    drop_last=False
-)
-
-rles, filename_and_class = test_with_tta(model, test_loader)
-
-preds = []
-for rle in rles[:len(CLASSES)]:
-    pred = decode_rle_to_mask(rle, height=2048, width=2048)
-    preds.append(pred)
-
-preds = np.stack(preds, 0)
+    preds = np.stack(preds, 0)
 
 
-classes, filename = zip(*[x.split("_") for x in filename_and_class])
+    classes, filename = zip(*[x.split("_") for x in filename_and_class])
 
-image_name = [os.path.basename(f) for f in filename]
+    image_name = [os.path.basename(f) for f in filename]
 
-df = pd.DataFrame({
-    "image_name": image_name,
-    "class": classes,
-    "rle": rles,
-})
+    df = pd.DataFrame({
+        "image_name": image_name,
+        "class": classes,
+        "rle": rles,
+    })
 
-df.to_csv("unetpp_base.csv", index=False)
+    df.to_csv("unetpp_base.csv", index=False)
 
-df["image_name"].nunique()
+    df["image_name"].nunique()
+
+
+if __name__ == "__main__":
+    main()
